pyproject.toml
[project]
name = "ai-mock-interviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "langchain[mistralai]>=0.3.27",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
